[FIX] toolbar 분리
